Refactor ExerciseModal
