+++ src/App.js
@@ -8,9 +8,6 @@ import './react-calendar-heatmap.css';
 import ExerciseButtonList from './components/ExerciseButtonList.js';
 import Calendar from './components/Calendar.js';
 import ExerciseModal from './components/ExerciseModal.js';
-import RandomWordsList from './resources/RandomWordsList.js';
-import GenreList from './resources/GenreList.js';
-import TextTypeList from './resources/TextTypeList.js';
 
 import Header from './components/Header.js'
 
@@ -19,9 +16,6 @@ const store = createStore(reducers, window.__REDUX_DEVTOOLS_EXTENSION__ && windo
 class App extends Component {
 
   state = {
-    randomWordsList: RandomWordsList,
-    genreList: GenreList,
-    textTypeList: TextTypeList,
     exercisesDone: [
       {
       task: "task",
@@ -42,50 +36,6 @@ class App extends Component {
     dateCount: [],
     currentExercise: "",
     pendingExerciseText: ""
-  }
-
-
-  handleCloseExercise = (e) => {
-    e.preventDefault();
-    this.setState({ 
-      showExerciseModal: false,
-      pendingExerciseText: "" 
-    });
-  }
-
-  handleShowExercise = (e) => {
-    e.preventDefault();
-    this.setState({ showExerciseModal: true });
-  }
-
-  getThreeWordsExercise = (e) => {
-    e.preventDefault();
-    let indeces = [];
-    let i = 0;
-    while (indeces.length < 3) { // for (let i = 0; i < 3; i++)
-      indeces[i] = Math.floor(Math.random() * RandomWordsList.length);
-      i++;
-      indeces = Array.from(new Set(indeces));
-    }
-    const words = RandomWordsList.filter(word => indeces.includes(RandomWordsList.indexOf(word)));
-    this.setState({
-      currentExercise: `Write a story including the words ${words[0]}, ${words[1]} and ${words[2]}.`,
-      showExerciseModal: true
-    });
-    console.log("Three word");
-    console.log(this.state);
-  }
-
-  getRandomExercise = (e) => {
-    e.preventDefault();
-    const indexGenre = Math.floor(Math.random() * GenreList.length);
-    const indexTextType = Math.floor(Math.random() * TextTypeList.length);
-    this.setState({
-      currentExercise: `Write a ${GenreList[indexGenre]} ${TextTypeList[indexTextType]}.`,
-      showExerciseModal: true
-    });
-    console.log("Random");
-    console.log(this.state);
   }
 
   handleDoingExercise = (e) => {
@@ -149,7 +99,6 @@ class App extends Component {
         <div className="app">
           <Header />
           <ExerciseModal /* show and handleClose is provided by redux */
-            exercise={this.state.currentExercise}
             pendingExerciseText={this.pendingExerciseText}
             handleDoingExercise={this.handleDoingExercise}
             handleSaveExercise={this.handleSaveExercise}
