+++ src/components/ExerciseButtonList.js
@@ -29,6 +29,7 @@ const mapDispatchToProps = dispatch => {
             dispatch(showExercise("random"));
         },
         openThreeWordExercise: () => {
+            console.log("open is called");
             dispatch(showExercise("three-word"));
         }
     };
