+++ src/ducks/handleExerciseText.js
@@ -1,3 +1,8 @@
-
+// ACTION TYPES
 export const HANDLE_DOING_EXERCISE = 'exercise/HANDLE_DOING_EXERCISE';
 export const SAVE_EXERCISE = 'exercise/SAVE_EXERCISE';
+
+// REDUCERS
+
+
+// ACTION CREATORS
